refactor halo for new ui
