+++ Cargo.toml
@@ -1,14 +1,4 @@
-[package]
-name = "halo"
-version = "0.1.0"
-edition = "2021"
-
-[dependencies]
-rusty_link = "0.4.3"
-artnet_protocol = "0.4.2"
-anyhow = "1.0.86"
-log = "0.4.22"
-midir = "0.10.1"
-clap = { version = "4.5.23", features = ["derive"] }
-crossterm = "0.28.1"
-eframe = "0.31.0"
+[workspace]
+members = ["crates/core", "crates/fixtures", "crates/halo", "crates/ui"]
+default-members = ["crates/halo"]
+resolver = "2"
