+++ crates/core/Cargo.toml
@@ -1,0 +1,14 @@
+[package]
+name = "halo-core"
+version = "0.1.0"
+authors = ["Rob Morgan <robbym@gmail.com>"]
+edition = "2021"
+
+[dependencies]
+halo-fixtures = { path = "../fixtures" }
+rusty_link = "0.4.3"
+artnet_protocol = "0.4.2"
+anyhow = "1.0.86"
+log = "0.4.22"
+crossterm = "0.28.1"
+midir = "0.10.1"
