+++ crates/core/src/midi/mod.rs
@@ -1,0 +1,1 @@
+pub mod midi;
