+++ crates/fixtures/src/lib.rs
@@ -1,4 +1,3 @@
-use crate::channel_layout;
 use std::collections::HashMap;
 
 #[derive(Clone, Debug)]
