+++ crates/halo/Cargo.toml
@@ -1,0 +1,29 @@
+# [lib]
+# name = "halo"
+# path = "src/lib.rs"
+# bench = false
+
+[[bin]]
+name = "halo"
+path = "src/main.rs"
+bench = false
+
+[package]
+name = "halo"
+description = "Realtime lighting console for live performances and precise automation."
+version = "0.1.0"
+authors = ["Rob Morgan <robbym@gmail.com>"]
+edition = "2021"
+
+[dependencies]
+halo-core = { path = "../core" }
+halo-ui = { path = "../ui" }
+halo-fixtures = { path = "../fixtures" }
+rusty_link = "0.4.3"
+artnet_protocol = "0.4.2"
+anyhow = "1.0.86"
+log = "0.4.22"
+midir = "0.10.1"
+clap = { version = "4.5.23", features = ["derive"] }
+crossterm = "0.28.1"
+eframe = "0.31.0"
