+++ crates/halo/src/mod.rs
@@ -1,0 +1,14 @@
+mod effects;
+mod fixtures;
+mod ui;
+
+mod ableton_link;
+mod artnet;
+mod console;
+mod cue;
+mod effect;
+mod fixture;
+mod midi;
+mod rhythm;
+mod ui;
+mod visualization;
